+++ examples/enc_raster.rs
@@ -1,5 +1,5 @@
 // Raster encoding example
-use gift::Encoder;
+use gift::{Encoder, Step};
 use pix::gray::Gray8;
 use pix::rgb::SRgb8;
 use pix::{Palette, Raster};
@@ -8,7 +8,7 @@ use std::fs::File;
 
 fn main() -> Result<(), Box<dyn Error>> {
     let mut f = File::create("enc_raster.gif")?;
-    let mut enc = Encoder::new(&mut f).into_raster_enc();
+    let mut enc = Encoder::new(&mut f).into_step_enc();
     let mut raster = Raster::with_clear(4, 4);
     *raster.pixel_mut(0, 0) = Gray8::new(1);
     *raster.pixel_mut(1, 1) = Gray8::new(1);
@@ -17,6 +17,7 @@ fn main() -> Result<(), Box<dyn Error>> {
     let mut palette = Palette::new(2);
     palette.set_entry(SRgb8::new(0xFF, 0, 0));
     palette.set_entry(SRgb8::new(0xFF, 0xFF, 0));
-    enc.encode_indexed_raster(&raster, &palette)?;
+    let step = Step::with_indexed(raster, palette);
+    enc.encode_step(&step)?;
     Ok(())
 }