+++ src/private.rs
@@ -3,20 +3,37 @@
 // Copyright (c) 2019-2020  Douglas Lau
 //
 //! Private module for top-level items
-use crate::{decode, encode, Result};
-use pix::rgb::SRgba8;
-use pix::Raster;
+use crate::{block::GraphicControl, decode, encode, Result};
+use pix::{gray::Gray8, rgb::SRgba8, Palette, Raster};
 use std::io::{BufReader, BufWriter, Read, Write};
+
+/// Raster for an animation step.
+pub(crate) enum StepRaster {
+    /// True color 24-bit raster
+    TrueColor(Raster<SRgba8>),
+    /// Indexed color 8-bit raster
+    Indexed(Raster<Gray8>, Palette),
+}
+
+/// One step of an animation.
+pub struct Step {
+    /// Raster of the animation step
+    pub(crate) raster: StepRaster,
+    /// Graphic control for the step
+    pub(crate) graphic_control_ext: Option<GraphicControl>,
+}
 
 /// GIF file decoder
 ///
 /// Can be converted to one of three `Iterator`s:
-/// * [into_iter] / [into_rasters] for high-level `Raster`s
+/// * [into_iter] / [into_steps] for high-level [Step]s
 /// * [into_frames] for mid-level [Frame]s
 /// * [into_blocks] for low-level [Block]s
 ///
 /// ## Example: Get a `Raster` from a GIF
 /// ```
+/// use gift::Decoder;
+///
 /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
 /// # let gif = &[
 /// #   0x47, 0x49, 0x46, 0x38, 0x39, 0x61, 0x02, 0x00,
@@ -26,9 +43,10 @@ use std::io::{BufReader, BufWriter, Read, Write};
 /// #   0x10, 0x05, 0x00, 0x3b,
 /// # ][..];
 /// // ... open a `File` as "gif"
-/// if let Some(raster) = gift::Decoder::new(gif).into_iter().next() {
+/// if let Some(step) = Decoder::new(gif).into_steps().next() {
 ///     // was there a decoding error?
-///     let raster = raster?;
+///     let step = step?;
+///     let raster = step.raster();
 ///     // ... work with raster
 /// }
 /// # Ok(())
@@ -40,13 +58,77 @@ use std::io::{BufReader, BufWriter, Read, Write};
 /// [into_blocks]: struct.Decoder.html#method.into_blocks
 /// [into_frames]: struct.Decoder.html#method.into_frames
 /// [into_iter]: struct.Decoder.html#method.into_iter
-/// [into_rasters]: struct.Decoder.html#method.into_rasters
+/// [into_steps]: struct.Decoder.html#method.into_steps
+/// [Step]: struct.Step.html
 ///
 pub struct Decoder<R: Read> {
     /// Reader for input data
     reader: R,
     /// Maximum image size, in bytes
     max_image_sz: Option<usize>,
+}
+
+impl Step {
+    /// Create an animation step with a true color raster.
+    pub fn with_true_color(raster: Raster<SRgba8>) -> Self {
+        let raster = StepRaster::TrueColor(raster);
+        Step {
+            raster,
+            graphic_control_ext: None,
+        }
+    }
+
+    /// Create an animation step with an indexed raster.
+    pub fn with_indexed(raster: Raster<Gray8>, palette: Palette) -> Self {
+        let raster = StepRaster::Indexed(raster, palette);
+        Step {
+            raster,
+            graphic_control_ext: None,
+        }
+    }
+
+    /// Adjust the transparent color.
+    pub(crate) fn with_transparent_color(mut self, clr: Option<u8>) -> Self {
+        if clr.is_some() || self.graphic_control_ext.is_some() {
+            let mut control = self.graphic_control_ext.unwrap_or_default();
+            control.set_transparent_color(clr);
+            if control != GraphicControl::default() {
+                self.graphic_control_ext = Some(control);
+            } else {
+                self.graphic_control_ext = None;
+            }
+        }
+        self
+    }
+
+    /// Adjust the delay time.
+    pub fn with_delay_time_cs(mut self, delay: Option<u16>) -> Self {
+        if delay.is_some() || self.graphic_control_ext.is_some() {
+            let mut control = self.graphic_control_ext.unwrap_or_default();
+            if let Some(delay) = delay {
+                control.set_delay_time_cs(delay);
+            }
+            if control != GraphicControl::default() {
+                self.graphic_control_ext = Some(control);
+            } else {
+                self.graphic_control_ext = None;
+            }
+        }
+        self
+    }
+
+    /// Get the raster
+    pub fn raster(&self) -> &Raster<SRgba8> {
+        match &self.raster {
+            StepRaster::TrueColor(ref r) => r,
+            StepRaster::Indexed(_, _) => todo!("convert to true color"),
+        }
+    }
+
+    /// Get the delay time in centiseconds
+    pub fn delay_time_cs(&self) -> Option<u16> {
+        self.graphic_control_ext.map(|c| c.delay_time_cs())
+    }
 }
 
 impl<R: Read> Decoder<BufReader<R>> {
@@ -81,38 +163,38 @@ impl<R: Read> Decoder<R> {
         decode::Frames::new(self.into_blocks())
     }
 
-    /// Convert into a raster `Iterator`.
-    pub fn into_rasters(self) -> decode::Rasters<R> {
-        decode::Rasters::new(self.into_frames())
+    /// Convert into a step `Iterator`.
+    pub fn into_steps(self) -> decode::Steps<R> {
+        decode::Steps::new(self.into_frames())
     }
 }
 
 impl<R: Read> IntoIterator for Decoder<R> {
-    type Item = Result<Raster<SRgba8>>;
-    type IntoIter = decode::Rasters<R>;
+    type Item = Result<Step>;
+    type IntoIter = decode::Steps<R>;
 
-    /// Convert into a raster `Iterator`
+    /// Convert into a raster step `Iterator`
     fn into_iter(self) -> Self::IntoIter {
-        self.into_rasters()
+        self.into_steps()
     }
 }
 
 /// GIF file encoder
 ///
 /// Can be converted to one of three encoders:
-/// * [into_raster_enc] for high-level `Raster`s
+/// * [into_step_enc] for high-level [Step]s
 /// * [into_frame_enc] for mid-level [Frame]s
 /// * [into_block_enc] for low-level [Block]s
 ///
 /// ## Encoding Example
 /// ```
-/// use gift::Encoder;
+/// use gift::{Encoder, Step};
 /// use pix::{gray::Gray8, Palette, Raster, rgb::SRgb8};
 /// use std::error::Error;
 /// use std::io::Write;
 ///
 /// fn encode<W: Write>(mut w: W) -> Result<(), Box<dyn Error>> {
-///     let mut enc = Encoder::new(&mut w).into_raster_enc();
+///     let mut enc = Encoder::new(&mut w).into_step_enc();
 ///     let mut raster = Raster::with_clear(4, 4);
 ///     *raster.pixel_mut(0, 0) = Gray8::new(1);
 ///     *raster.pixel_mut(1, 1) = Gray8::new(1);
@@ -121,7 +203,8 @@ impl<R: Read> IntoIterator for Decoder<R> {
 ///     let mut palette = Palette::new(2);
 ///     palette.set_entry(SRgb8::new(0xFF, 0, 0));
 ///     palette.set_entry(SRgb8::new(0xFF, 0xFF, 0));
-///     enc.encode_indexed_raster(&raster, &palette)?;
+///     let step = Step::with_indexed(raster, palette);
+///     enc.encode_step(&step)?;
 ///     Ok(())
 /// }
 /// ```
@@ -130,7 +213,8 @@ impl<R: Read> IntoIterator for Decoder<R> {
 /// [Frame]: block/struct.Frame.html
 /// [into_block_enc]: struct.Encoder.html#method.into_block_enc
 /// [into_frame_enc]: struct.Encoder.html#method.into_frame_enc
-/// [into_raster_enc]: struct.Encoder.html#method.into_raster_enc
+/// [into_step_enc]: struct.Encoder.html#method.into_step_enc
+/// [Step]: struct.Step.html
 pub struct Encoder<W: Write> {
     /// Writer for output data
     writer: W,
@@ -159,8 +243,8 @@ impl<W: Write> Encoder<W> {
         encode::FrameEnc::new(self.into_block_enc())
     }
 
-    /// Convert into a raster encoder.
-    pub fn into_raster_enc(self) -> encode::RasterEnc<W> {
-        encode::RasterEnc::new(self.into_frame_enc())
+    /// Convert into a step encoder.
+    pub fn into_step_enc(self) -> encode::StepEnc<W> {
+        encode::StepEnc::new(self.into_frame_enc())
     }
 }